Refactor: Use `kotlin.io.encoding.Base64` and relocate crypto qualifiers
- Replaced `java.util.Base64` with `kotlin.io.encoding.Base64` in `KeyAttestationViewModel` for Base64 encoding/decoding operations.
- Moved Hilt qualifiers (`EC`, `ECDH`, `RSA`) from `dev.keiji.deviceintegrity.di.qualifier` to `dev.keiji.deviceintegrity.crypto.contract.qualifier`.
- Updated import paths for these qualifiers in `KeyAttestationViewModel` and `CryptoModule`.
- Added `javax.inject` dependency to `crypto:contract` and `crypto:impl` modules.
- Added `crypto:contract` as a dependency to `ui:main`.
- Simplified `KeyAttestationVerifyApiClient` by removing unused imports.
- Made `CryptoModule` an object.
