+++ android/crypto/contract/build.gradle.kts
@@ -31,4 +31,5 @@ android {
 }
 
 dependencies {
+    implementation(libs.javax.inject)
 }
