+++ android/ui/main/src/main/java/dev/keiji/deviceintegrity/ui/main/keyattestation/KeyAttestationViewModel.kt
@@ -2,24 +2,24 @@ package dev.keiji.deviceintegrity.ui.main.keyattestation
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import java.util.Base64 // Replaced android.util.Base64
+import dagger.hilt.android.lifecycle.HiltViewModel
 import dev.keiji.deviceintegrity.api.keyattestation.KeyAttestationVerifyApiClient
 import dev.keiji.deviceintegrity.api.keyattestation.PrepareRequest
 import dev.keiji.deviceintegrity.api.keyattestation.VerifyEcRequest
 import dev.keiji.deviceintegrity.crypto.contract.Signer
+import dev.keiji.deviceintegrity.crypto.contract.qualifier.EC
 import dev.keiji.deviceintegrity.repository.contract.KeyPairRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
-import kotlinx.coroutines.launch // Ensure this is imported
+import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import java.security.SecureRandom
 import java.util.UUID
 import javax.inject.Inject
-import dagger.hilt.android.lifecycle.HiltViewModel
-import dev.keiji.deviceintegrity.di.qualifier.EC
+import kotlin.io.encoding.Base64
 
 @HiltViewModel
 class KeyAttestationViewModel @Inject constructor(
@@ -49,7 +49,13 @@ class KeyAttestationViewModel @Inject constructor(
     // Action to fetch Nonce/Challenge
     fun fetchNonceChallenge() {
         viewModelScope.launch {
-            _uiState.update { it.copy(status = "Fetching Nonce/Challenge...", nonce = "", challenge = "") }
+            _uiState.update {
+                it.copy(
+                    status = "Fetching Nonce/Challenge...",
+                    nonce = "",
+                    challenge = ""
+                )
+            }
             try {
                 val newSessionId = UUID.randomUUID().toString()
                 _uiState.update { it.copy(sessionId = newSessionId) }
@@ -83,7 +89,12 @@ class KeyAttestationViewModel @Inject constructor(
     // Action to generate KeyPair
     fun generateKeyPair() {
         viewModelScope.launch {
-            _uiState.update { it.copy(status = "Generating KeyPair...", generatedKeyPairData = null) }
+            _uiState.update {
+                it.copy(
+                    status = "Generating KeyPair...",
+                    generatedKeyPairData = null
+                )
+            }
 
             val currentChallenge = uiState.value.challenge
             if (currentChallenge.isEmpty()) {
@@ -95,7 +106,7 @@ class KeyAttestationViewModel @Inject constructor(
                 // Decode the challenge from Base64Url
                 // URL_SAFE is used because the server sends it as Base64UrlEncoded.
                 val decodedChallenge = withContext(Dispatchers.Default) {
-                    Base64.getUrlDecoder().decode(currentChallenge)
+                    Base64.UrlSafe.decode(currentChallenge)
                 }
 
                 // Perform key generation on IO dispatcher
@@ -129,7 +140,8 @@ class KeyAttestationViewModel @Inject constructor(
                 _uiState.update { it.copy(status = "SessionId is missing. Fetch Nonce/Challenge first.") }
                 return@launch
             }
-            if (currentKeyPairData?.keyPair == null) {
+            val keyPair = currentKeyPairData?.keyPair
+            if (keyPair == null) {
                 _uiState.update { it.copy(status = "KeyPair is not generated yet.") }
                 return@launch
             }
@@ -146,20 +158,21 @@ class KeyAttestationViewModel @Inject constructor(
                     SecureRandom().nextBytes(nonceB)
 
                     // Server nonce is Base64URL Encoded
-                    val decodedServerNonce = Base64.getUrlDecoder().decode(serverNonceB64Url)
+                    val decodedServerNonce = Base64.UrlSafe.decode(serverNonceB64Url)
                     val dataToSign = decodedServerNonce + nonceB
 
                     // 2. Signing
-                    val privateKey = currentKeyPairData.keyPair.private
+                    val privateKey = keyPair.private
                     val signatureData = signer.sign(dataToSign, privateKey)
 
                     // 3. Encoding for Request (Base64URL as per task, without padding)
-                    val urlSafeEncoder = Base64.getUrlEncoder().withoutPadding()
-                    val signedDataBase64UrlEncoded = urlSafeEncoder.encodeToString(signatureData)
-                    val nonceBBase64UrlEncoded = urlSafeEncoder.encodeToString(nonceB)
-                    val certificateChainBase64UrlEncoded = currentKeyPairData.certificates.map { cert ->
-                        urlSafeEncoder.encodeToString(cert.encoded)
-                    }
+                    val urlSafeEncoder = Base64.UrlSafe
+                    val signedDataBase64UrlEncoded = urlSafeEncoder.encode(signatureData)
+                    val nonceBBase64UrlEncoded = urlSafeEncoder.encode(nonceB)
+                    val certificateChainBase64UrlEncoded =
+                        currentKeyPairData.certificates.map { cert ->
+                            urlSafeEncoder.encode(cert.encoded)
+                        }
 
                     // 4. API Call
                     val request = VerifyEcRequest(
